backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
nhl-api-py>=3.1.1
pandas>=2.0.0
numpy>=1.24.0